implement last_commit_hash check
