+++ crates/astria-sequencer/src/transaction/mod.rs
@@ -73,7 +73,6 @@ impl ActionHandler for UnsignedTransaction {
             "invalid nonce, tx nonce must match account nonce"
         );
 
-        // do we need to make a StateDelta here so we can check the actions on the successive state?
         for action in &self.actions {
             match action {
                 Action::Transfer(act) => act
